okta-automation-toolkit/requirements.txt
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
idna==3.11
pycparser==3.0
PyJWT==2.11.0
requests==2.32.5
urllib3==2.6.3
